Refactor notifs ; Implement a working notifications system ; Still needs to remove the TODO / add tests
