+++ app/src/main/java/com/github/onedirection/EventsView.java
@@ -8,7 +8,6 @@ import android.widget.TextView;
 
 import com.github.onedirection.events.Event;
 import com.github.onedirection.events.EventCreator;
-import com.github.onedirection.notifs.Notifications;
 
 import java.util.HashSet;
