+++ app/src/main/java/com/github/onedirection/navigation/NavigationActivity.java
@@ -19,6 +19,7 @@ import androidx.navigation.ui.NavigationUI;
 import com.github.onedirection.R;
 import com.github.onedirection.authentication.FirebaseAuthentication;
 import com.github.onedirection.events.EventCreator;
+import com.github.onedirection.notification.Notifications;
 import com.google.android.material.navigation.NavigationView;
 
 public class NavigationActivity extends AppCompatActivity {
@@ -65,6 +66,10 @@ public class NavigationActivity extends AppCompatActivity {
                 return false;
             }
         });
+
+        // Starts the notifications handler.
+        // DO NOT REMOVE THIS LINE otherwise notifications wont work!
+        Notifications.getInstance(getApplicationContext());
     }
 
     @Override
