+++ app/src/main/java/com/github/onedirection/notification/NotificationIdGenerator.java
@@ -1,4 +1,4 @@
-package com.github.onedirection.notifs;
+package com.github.onedirection.notification;
 
 public class NotificationIdGenerator {
     private static Integer counter = 0;
